+++ dapps/marketplace/src/pages/listings/Listings.js
@@ -11,7 +11,6 @@ import withGrowthCampaign from 'hoc/withGrowthCampaign'
 import withGrowthRewards from 'hoc/withGrowthRewards'
 import withWallet from 'hoc/withWallet'
 import withTokenBalance from 'hoc/withTokenBalance'
-import withIsMobile from 'hoc/withIsMobile'
 
 import BottomScrollListener from 'components/BottomScrollListener'
 import QueryError from 'components/QueryError'
@@ -149,9 +148,6 @@ class Listings extends Component {
       <>
         <DocumentTitle pageTitle={<fbt desc="listings.title">Listings</fbt>} />
         <div className="container listings-container">
-          {this.props.isMobile ? (
-            <Search className="search" placeholder />
-          ) : null}
           <Query
             query={query}
             variables={vars}
@@ -281,7 +277,7 @@ class Listings extends Component {
 
 export default withGrowthRewards(
   withGrowthCampaign(
-    withWallet(withTokenBalance(withCreatorConfig(withIsMobile(Listings)))),
+    withWallet(withTokenBalance(withCreatorConfig(Listings))),
     {
       fetchPolicy: 'cache-first',
       queryEvenIfNotEnrolled: true,
